Move NormalizeBranchName to Utilities
