+++ src/Maestro/Maestro.Data/Models/Repository.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.DotNet.DarcLib;
+using Microsoft.DotNet.Services.Utilities;
 using Newtonsoft.Json;
 
 namespace Maestro.Data.Models
@@ -98,11 +99,11 @@ namespace Maestro.Data.Models
         {
             get
             {
-                return IGitRepoExtension.NormalizeBranchName(_branch);
+                return Utilities.NormalizeBranchName(_branch);
             }
             set
             {
-                _branch = IGitRepoExtension.NormalizeBranchName(value);
+                _branch = Utilities.NormalizeBranchName(value);
             }
         }
