+++ src/Shared/Microsoft.DotNet.Services.Utilities/Microsoft.DotNet.Services.Utilities.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <RuntimeIdentifier>win7-x64</RuntimeIdentifier>
+    <TargetLatestRuntimePatch>False</TargetLatestRuntimePatch>
+    <SignAssembly>false</SignAssembly>
+    <LangVersion>7.1</LangVersion>
+    <RootNamespace>Microsoft.DotNet.Services.Utilities</RootNamespace>
+    <AssemblyName>Microsoft.DotNet.Services.Utilities</AssemblyName>
+  </PropertyGroup>
+
+</Project>
